// apps/api/src/dionysus/metadata/certifications/types/certification.ts
import { CertificationType } from "@ncfritz/olympus-model";
import { Timestamped, Wrapped } from "../../types/metadata";

export type GraphQlCertification = {
  certification: string;
  country: string;
  createdTime: string;
  lastUpdatedTime: string;
  meaning: string;
  order: number;
  type: CertificationType;
};

export type GraphQlCertificationWrapper = Timestamped &
  Wrapped<GraphQlCertification, "certification">;